+++ blocklets/aigne-runtime/CHANGELOG.md
@@ -1,5 +1,31 @@
 # Changelog
 
+## [1.4.18](https://github.com/AIGNE-io/aigne-framework/compare/aigne-runtime-v1.4.17...aigne-runtime-v1.4.18) (2025-09-18)
+
+
+### Bug Fixes
+
+* correct flat import, remove default import error ([#519](https://github.com/AIGNE-io/aigne-framework/issues/519)) ([6e8fd30](https://github.com/AIGNE-io/aigne-framework/commit/6e8fd30b89be1300ab6214f13fd273e51e25329b))
+* correct flat import, remove default import error ([#519](https://github.com/AIGNE-io/aigne-framework/issues/519)) ([6e8fd30](https://github.com/AIGNE-io/aigne-framework/commit/6e8fd30b89be1300ab6214f13fd273e51e25329b))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/aigne-hub bumped to 0.9.6
+
+## [1.4.17](https://github.com/AIGNE-io/aigne-framework/compare/aigne-runtime-v1.4.16...aigne-runtime-v1.4.17) (2025-09-18)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/aigne-hub bumped to 0.9.5
+    * @aigne/core bumped to 1.60.3
+    * @aigne/transport bumped to 0.15.8
+
 ## [1.4.16](https://github.com/AIGNE-io/aigne-framework/compare/aigne-runtime-v1.4.15...aigne-runtime-v1.4.16) (2025-09-16)