+++ models/bedrock/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [0.10.4](https://github.com/AIGNE-io/aigne-framework/compare/bedrock-v0.10.3...bedrock-v0.10.4) (2025-09-18)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/core bumped to 1.60.3
+  * devDependencies
+    * @aigne/test-utils bumped to 0.5.52
+
 ## [0.10.3](https://github.com/AIGNE-io/aigne-framework/compare/bedrock-v0.10.2...bedrock-v0.10.3) (2025-09-11)