+++ models/lmstudio/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [1.0.14](https://github.com/AIGNE-io/aigne-framework/compare/lmstudio-v1.0.13...lmstudio-v1.0.14) (2025-09-18)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/openai bumped to 0.15.4
+  * devDependencies
+    * @aigne/core bumped to 1.60.3
+    * @aigne/test-utils bumped to 0.5.52
+
 ## [1.0.13](https://github.com/AIGNE-io/aigne-framework/compare/lmstudio-v1.0.12...lmstudio-v1.0.13) (2025-09-11)