+++ packages/test-utils/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [0.5.52](https://github.com/AIGNE-io/aigne-framework/compare/test-utils-v0.5.51...test-utils-v0.5.52) (2025-09-18)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/core bumped to 1.60.3
+
 ## [0.5.51](https://github.com/AIGNE-io/aigne-framework/compare/test-utils-v0.5.50...test-utils-v0.5.51) (2025-09-11)